+++ src/guess-number.js
@@ -1,8 +1,14 @@
 
 
-button  {
-    height: 100px;
-    width: 100px;
-    background-color: blue
+const guess-number = Input
+
+
+function guess-number(text){
+  );
+    if(guess - expected = 0) 
+    {console.log('Congratulations!');}
+    else{console.log('Wrong!');}
 
 }
+
+export default guess-number.js;
